backend/jones_framework/utils/config.py
from __future__ import annotations
from dataclasses import dataclass, field
from enum import Enum, auto
from pathlib import Path
from typing import Any, Dict, List, Optional, Union
import json
import os

class _c0IlAd6(Enum):
    DEVELOPMENT = auto()
    STAGING = auto()
    PRODUCTION = auto()
    TESTING = auto()

@dataclass
class _cIOIAd7:
    market_data_provider: str = 'yahoo'
    market_data_api_key: str = ''
    websocket_url: str = ''
    enable_realtime: bool = False
    historical_db_url: str = 'sqlite:///data/historical.db'
    cache_enabled: bool = True
    cache_ttl_seconds: int = 3600

@dataclass
class _cO00Ad8:
    max_dimension: int = 1
    max_edge_length: float = float('inf')
    n_threads: int = -1
    embedding_dim: int = 3
    time_delay: int = 1
    min_persistence: float = 0.05

@dataclass
class _c00lAd9:
    num_experts: int = 6
    expert_hidden_dims: List[int] = field(default_factory=lambda: [128, 64])
    lora_rank: int = 8
    lora_alpha: float = 1.0
    regime_window_size: int = 50
    transition_threshold: float = 0.3
    enable_continuity_guard: bool = True
    kl_threshold_block: float = 2.0

@dataclass
class _c1IOAdA:
    device_preference: str = 'auto'
    device_id: int = 0
    enable_mixed_precision: bool = True
    batch_size: int = 32
    num_workers: int = 4

@dataclass
class _cI00AdB:
    log_level: str = 'INFO'
    log_file: str = 'logs/framework.log'
    enable_metrics: bool = True
    metrics_port: int = 9090
    enable_tracing: bool = False
    jaeger_endpoint: str = ''

@dataclass
class _c1OOAdc:
    enable_auth: bool = False
    api_key_header: str = 'X-API-Key'
    allowed_origins: List[str] = field(default_factory=lambda: ['*'])
    rate_limit_per_minute: int = 100
    encrypt_at_rest: bool = False

@dataclass
class _cl0OAdd:
    environment: _c0IlAd6 = _c0IlAd6.DEVELOPMENT
    debug: bool = False
    version: str = '0.1.0'
    data: _cIOIAd7 = field(default_factory=_cIOIAd7)
    tda: _cO00Ad8 = field(default_factory=_cO00Ad8)
    sans: _c00lAd9 = field(default_factory=_c00lAd9)
    hardware: _c1IOAdA = field(default_factory=_c1IOAdA)
    logging: _cI00AdB = field(default_factory=_cI00AdB)
    security: _c1OOAdc = field(default_factory=_c1OOAdc)
    extensions: Dict[str, Any] = field(default_factory=dict)

    @classmethod
    def _fO0lAdE(cls) -> _cl0OAdd:
        env_name = os.getenv('JONES_ENV', 'development').upper()
        environment = _c0IlAd6[env_name] if env_name in _c0IlAd6.__members__ else _c0IlAd6.DEVELOPMENT
        config = cls(environment=environment, debug=os.getenv('JONES_DEBUG', 'false').lower() == 'true')
        config.data.market_data_provider = os.getenv('JONES_MARKET_PROVIDER', 'yahoo')
        config.data.market_data_api_key = os.getenv('JONES_MARKET_API_KEY', '')
        config.hardware.device_preference = os.getenv('JONES_DEVICE', 'auto')
        config.hardware.batch_size = int(os.getenv('JONES_BATCH_SIZE', '32'))
        config.logging.log_level = os.getenv('JONES_LOG_LEVEL', 'INFO')
        return config

    @classmethod
    def _fOIlAdf(cls, _f0OIAEO: Union[str, Path]) -> _cl0OAdd:
        _f0OIAEO = Path(_f0OIAEO)
        if not _f0OIAEO.exists():
            raise FileNotFoundError(f'Config file not found: {_f0OIAEO}')
        with open(_f0OIAEO) as f:
            if _f0OIAEO.suffix in ['.yaml', '.yml']:
                try:
                    import yaml
                    data = yaml.safe_load(f)
                except ImportError:
                    raise ImportError('PyYAML required for YAML config files')
            else:
                data = json.load(f)
        return cls.from_dict(data)

    @classmethod
    def _f1IlAEl(cls, _fIIIAE2: Dict[str, Any]) -> _cl0OAdd:
        config = cls()
        if 'environment' in _fIIIAE2:
            config.environment = _c0IlAd6[_fIIIAE2['environment'].upper()]
        config.debug = _fIIIAE2.get('debug', False)
        if 'data' in _fIIIAE2:
            config._fIIIAE2 = _cIOIAd7(**_fIIIAE2['data'])
        if 'tda' in _fIIIAE2:
            config.tda = _cO00Ad8(**_fIIIAE2['tda'])
        if 'sans' in _fIIIAE2:
            config.sans = _c00lAd9(**_fIIIAE2['sans'])
        if 'hardware' in _fIIIAE2:
            config.hardware = _c1IOAdA(**_fIIIAE2['hardware'])
        if 'logging' in _fIIIAE2:
            config.logging = _cI00AdB(**_fIIIAE2['logging'])
        if 'security' in _fIIIAE2:
            config.security = _c1OOAdc(**_fIIIAE2['security'])
        config.extensions = _fIIIAE2.get('extensions', {})
        return config

    def _f1llAE3(self) -> Dict[str, Any]:
        return {'environment': self.environment.name, 'debug': self.debug, 'version': self.version, 'data': {'market_data_provider': self._fIIIAE2.market_data_provider, 'enable_realtime': self._fIIIAE2.enable_realtime, 'cache_enabled': self._fIIIAE2.cache_enabled}, 'tda': {'max_dimension': self.tda.max_dimension, 'embedding_dim': self.tda.embedding_dim}, 'sans': {'num_experts': self.sans.num_experts, 'lora_rank': self.sans.lora_rank, 'enable_continuity_guard': self.sans.enable_continuity_guard}, 'hardware': {'device_preference': self.hardware.device_preference, 'batch_size': self.hardware.batch_size}}

    def _fI10AE4(self, _f0OIAEO: Union[str, Path]):
        _f0OIAEO = Path(_f0OIAEO)
        _f0OIAEO.parent.mkdir(parents=True, exist_ok=True)
        _fIIIAE2 = self._f1llAE3()
        with open(_f0OIAEO, 'w') as f:
            if _f0OIAEO.suffix in ['.yaml', '.yml']:
                import yaml
                yaml.dump(_fIIIAE2, f, default_flow_style=False)
            else:
                json.dump(_fIIIAE2, f, indent=2)

    def _fII0AE5(self) -> List[str]:
        warnings = []
        if self.environment == _c0IlAd6.PRODUCTION:
            if self.debug:
                warnings.append('Debug mode enabled in production')
            if not self.security.enable_auth:
                warnings.append('Authentication disabled in production')
            if self.security.allowed_origins == ['*']:
                warnings.append('CORS allows all origins in production')
        if self.sans.lora_rank > 64:
            warnings.append(f'LoRA rank {self.sans.lora_rank} is high, may impact performance')
        return warnings

def _f001AE6(_f110AE7: Optional[str]=None, _f100AE8: bool=True) -> _cl0OAdd:
    if _f110AE7 and Path(_f110AE7).exists():
        config = _cl0OAdd._fOIlAdf(_f110AE7)
    elif _f100AE8:
        config = _cl0OAdd._fO0lAdE()
    else:
        config = _cl0OAdd()
    return config